Rather than having bindings.utils._get_handle.pyx depend on driver and runtime and define the getters there, this flips things so driver and runtime register their own handlers.

While this doesn't resolve all cyclical imports as required by #789, this was a fairly straightforward change that improves import times by about 10%:

- Cold import time of bindings.cuda.nvrtc before this change: 59.1 ms +- 1.7 ms
- Cold import time of bindings.cuda.nvrtc after this change: 53.8 ms +- 2.4 ms
- Baseline Python startup: 8.36 ms +- 0.31 ms

(This was measured using [this pyperf benchmark](https://gist.github.com/mdboom/4a5152e4ab247f2d5a662a7bd1c1883d), starting with a fresh Python process each time.)

This is generated based on changes to cython_gen submitted in the private repo: https://github.com/NVIDIA/cuda-python-private/pull/131