@leofang 2025-08-19
> > @leofang: As far as workflow goes, should I merge this now or wait until the cython-gen changes that generated them are approved also?
> 
> We often merge here before merging the cython-gen changes. I'd say go ahead.
> 
> It's ideal to merge the cython-gen changes shortly before or after. I think you can even merge them without approval, although I generally feel most comfortable if I get an explicit approval there, too.

Thanks, Mike! Getting approval is always good. I did not merge last week because I wanted to push out the patch release first, and we can't keep asking Peter to re-spin the test every time we update the binding code. We should keep the changes minimal. After we pushed it out this morning, we are un-freezed and resume merging. 

Keep in mind that we haven't backported #835 to the codegens, so right now after regenerating the output will deviate from what's in the public. I am very occupied today and likely tomorrow due to the remaining CuPy release works, I hope we can get it done soon.